Workaround for single line comments
GitHub's API doesn't allow for `start_line` and `line` to be the same
number. For now we are working around this by making the `start_line`
one line earlier. We should be able to omit `start_line` and
`start_side` in the request body, but I am seeing API errors if they are
set to null.